BACKEND
- Changed how the route GET /get-todays-data generates data to send (some countries don't update covid data everyday);

- Added memoize to the generated data to prevent overload;

- Now everytime when covid data is updated, a new table called most_recent data is created with the most recent data from each country;

- Improved performance.

FRONTEND

- Now the table also displays the date when the data was inserted, HDI and Population density.

+++ backend/covidmap/models.py
@@ -54,7 +54,7 @@ class CovidData(models.Model):
                                                null=0
                                             )
     
-    population_density = models.BigIntegerField(verbose_name="Population Density",
+    population_density = models.FloatField(verbose_name="Population Density",
                                                 null=0
                                              )
     
@@ -78,6 +78,9 @@ class CovidData(models.Model):
     def continent_name(self):
         return self.continent.continent
     
+
+    
+
